+++ lib/src/http_result.dart
@@ -1,5 +1,4 @@
 import 'dart:io';
-
 import 'package:http/http.dart';
 
 class HttpResult<T> {
@@ -13,81 +12,19 @@ class HttpResult<T> {
       : _response = null,
         _data = null;
 
+  BaseRequest get request => (_response != null) ? _response.request : null;
+
   Response get response => _response;
 
   T get data => _data;
 
   dynamic get exception => _exception;
 
-  bool get isSuccess =>
-      (_response != null) && (status >= 200) && (status <= 299);
-
-  bool get isError => (_response != null) && (status >= 400);
-
-  bool get isOk => _is(HttpStatus.ok);
-
-  bool get isCreated => _is(HttpStatus.created);
-
-  bool get isAccepted => _is(HttpStatus.accepted);
-
-  bool get isNoContent => _is(HttpStatus.noContent);
-
-  bool get isBadRequest => _is(HttpStatus.badRequest);
-
-  bool get isUnauthorized => _is(HttpStatus.unauthorized);
-
-  bool get isForbidden => _is(HttpStatus.forbidden);
-
-  bool get isNotFound => _is(HttpStatus.notFound);
-
-  bool get isMethodNotAllowed => _is(HttpStatus.methodNotAllowed);
-
-  bool get isNotAcceptable => _is(HttpStatus.notAcceptable);
-
-  bool get isRequestTimeout => _is(HttpStatus.requestTimeout);
-
-  bool get isConflict => _is(HttpStatus.conflict);
-
-  bool get isGone => _is(HttpStatus.gone);
-
-  bool get isLengthRequired => _is(HttpStatus.lengthRequired);
-
-  bool get isPreconditionFailed => _is(HttpStatus.preconditionFailed);
-
-  bool get isRequestEntityTooLarge => _is(HttpStatus.requestEntityTooLarge);
-
-  bool get isUnsupportedMediaType => _is(HttpStatus.unsupportedMediaType);
-
-  bool get isExpectationFailed => _is(HttpStatus.expectationFailed);
-
-  bool get isLocked => _is(HttpStatus.locked);
-
-  bool get isFailedDependency => _is(HttpStatus.failedDependency);
-
-  bool get isPreconditionRequired => _is(HttpStatus.preconditionRequired);
-
-  bool get isTooManyRequests => _is(HttpStatus.tooManyRequests);
-
-  bool get isInternalServerError => _is(HttpStatus.internalServerError);
-
-  bool get isNotImplemented => _is(HttpStatus.notImplemented);
-
-  bool get isBadGateway => _is(HttpStatus.badGateway);
-
-  bool get isServiceUnavailable => _is(HttpStatus.serviceUnavailable);
-
-  bool get isGatewayTimeout => _is(HttpStatus.gatewayTimeout);
-
-  bool get isInternetConnection =>
-      hasException && (exception is SocketException);
-
-  bool hasStatus(int code) => _is(code);
-
-  bool get hasException => _exception != null;
+  bool hasStatus(int code) => status == code;
 
   bool get hasData => _data != null;
 
-  BaseRequest get request => (_response != null) ? _response.request : null;
+  bool get hasException => _exception != null;
 
   int get status => (_response != null) ? _response.statusCode : 0;
 
@@ -96,81 +33,144 @@ class HttpResult<T> {
   Map<String, String> get headers =>
       (_response != null) ? _response.headers : <String, String>{};
 
+  bool get success => (status >= 200) && (status <= 299);
+
+  bool get error => status >= 400;
+
+  bool get statusOk => hasStatus(HttpStatus.ok);
+
+  bool get statusCreated => hasStatus(HttpStatus.created);
+
+  bool get statusAccepted => hasStatus(HttpStatus.accepted);
+
+  bool get statusNoContent => hasStatus(HttpStatus.noContent);
+
+  bool get statusBadRequest => hasStatus(HttpStatus.badRequest);
+
+  bool get statusUnauthorized => hasStatus(HttpStatus.unauthorized);
+
+  bool get statusForbidden => hasStatus(HttpStatus.forbidden);
+
+  bool get statusNotFound => hasStatus(HttpStatus.notFound);
+
+  bool get statusMethodNotAllowed => hasStatus(HttpStatus.methodNotAllowed);
+
+  bool get statusNotAcceptable => hasStatus(HttpStatus.notAcceptable);
+
+  bool get statusRequestTimeout => hasStatus(HttpStatus.requestTimeout);
+
+  bool get statusConflict => hasStatus(HttpStatus.conflict);
+
+  bool get statusGone => hasStatus(HttpStatus.gone);
+
+  bool get statusLengthRequired => hasStatus(HttpStatus.lengthRequired);
+
+  bool get statusPreconditionFailed => hasStatus(HttpStatus.preconditionFailed);
+
+  bool get statusRequestEntityTooLarge =>
+      hasStatus(HttpStatus.requestEntityTooLarge);
+
+  bool get statusUnsupportedMediaType =>
+      hasStatus(HttpStatus.unsupportedMediaType);
+
+  bool get statusExpectationFailed => hasStatus(HttpStatus.expectationFailed);
+
+  bool get statusLocked => hasStatus(HttpStatus.locked);
+
+  bool get statusFailedDependency => hasStatus(HttpStatus.failedDependency);
+
+  bool get statusPreconditionRequired =>
+      hasStatus(HttpStatus.preconditionRequired);
+
+  bool get statusTooManyRequests => hasStatus(HttpStatus.tooManyRequests);
+
+  bool get statusInternalServerError =>
+      hasStatus(HttpStatus.internalServerError);
+
+  bool get statusNotImplemented => hasStatus(HttpStatus.notImplemented);
+
+  bool get statusBadGateway => hasStatus(HttpStatus.badGateway);
+
+  bool get statusServiceUnavailable => hasStatus(HttpStatus.serviceUnavailable);
+
+  bool get statusGatewayTimeout => hasStatus(HttpStatus.gatewayTimeout);
+
+  bool get statusInternetConnection =>
+      hasException && (exception is SocketException);
+
   void handle({
-    OnSuccess<T> success,
-    OnError error,
-    OnException exception,
-    OnFailure failure,
+    OnSuccess<T> onSuccess,
+    OnError onError,
+    OnException onException,
+    OnFailure onFailure,
   }) {
-    if (isSuccess) {
-      success?.call(_data, _response);
-    } else if (isError && (error != null)) {
-      error(_response);
-    } else if (hasException && (exception != null)) {
-      exception(_exception);
+    if (success) {
+      onSuccess?.call(_data, _response);
+    } else if (error && (onError != null)) {
+      onError(_response);
+    } else if (hasException && (onException != null)) {
+      onException(_exception);
     } else {
-      failure?.call(_response, _exception);
+      onFailure?.call(_response, _exception);
     }
   }
 
   void handleEmpty({
-    OnSuccessEmpty successful,
-    OnError error,
-    OnException exception,
-    OnFailure failure,
+    OnSuccessEmpty onSuccess,
+    OnError onError,
+    OnException onException,
+    OnFailure onFailure,
   }) {
-    if (isSuccess) {
-      successful?.call(_response);
-    } else if (isError && (error != null)) {
-      error(_response);
-    } else if (hasException && (exception != null)) {
-      exception(_exception);
+    if (success) {
+      onSuccess?.call(_response);
+    } else if (error && (onError != null)) {
+      onError(_response);
+    } else if (hasException && (onException != null)) {
+      onException(_exception);
     } else {
-      failure?.call(_response, _exception);
+      onFailure?.call(_response, _exception);
     }
   }
 
-  HttpResult<T> onSuccess(OnSuccess<T> success) {
-    if (isSuccess) {
-      success?.call(_data, _response);
-    }
-
-    return this;
-  }
-
-  HttpResult<T> onSuccessEmpty(OnSuccessEmpty success) {
-    if (isSuccess) {
-      success?.call(_response);
+  HttpResult<T> onSuccess(OnSuccess<T> onSuccess) {
+    if (success) {
+      onSuccess?.call(_data, _response);
     }
 
     return this;
   }
 
-  HttpResult<T> onError(OnError error) {
-    if (isError) {
-      error?.call(_response);
+  HttpResult<T> onSuccessEmpty(OnSuccessEmpty onSuccess) {
+    if (success) {
+      onSuccess?.call(_response);
     }
 
     return this;
   }
 
-  HttpResult<T> onException(OnException exception) {
+  HttpResult<T> onError(OnError onError) {
+    if (error) {
+      onError?.call(_response);
+    }
+
+    return this;
+  }
+
+  HttpResult<T> onException(OnException onException) {
     if (hasException) {
-      exception?.call(_exception);
+      onException?.call(_exception);
     }
 
     return this;
   }
 
-  HttpResult<T> onFailure(OnFailure failure) {
-    if (isError || hasException) {
-      failure?.call(_response, _exception);
+  HttpResult<T> onFailure(OnFailure onFailure) {
+    if (error || hasException) {
+      onFailure?.call(_response, _exception);
     }
 
     return this;
   }
-
-  bool _is(int code) => status == code;
 }
 
 typedef OnSuccess<T> = void Function(T data, Response response);